Remove MicdoodleCore dependency
Since the Galacticraft license is complicated to the point that I am
not sure anymore whether it is allowed or not to modify README.md,
I'm going to mangle the description of this useless fork here.

This is my second (or third, if one counts my unpublished and
unfinished attempt before @micdoodle8 quickly got his MC1.10 repo
going) take at making a relatively stable Galacticraft version for
the 1.10.2 version of the game before @micdoodle8 does it.

Aside from its complete uselessness and waste of some possibly better
spent time doing something else, it attempts to make the codebase
fancier and easier to comprehend by removing the MicdoodleCore coremod
dependency, fixing a random issue with arc lamps' light not getting
removed in the process.

Although it was not extensively tested, that modification probably
sacrifices a lot of compatibility with some purely visual mods.

Special thanks to @LexManos for the <rmuvd>personalized coremod
rant</rmuvd> major inspiration.